package restructuring
Added new activity for job intent service
