+++ app/src/main/AndroidManifest.xml
@@ -1,14 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:tools="http://schemas.android.com/tools" >
 
     <uses-permission android:name="android.permission.POST_NOTIFICATIONS" />
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
     <uses-permission android:name="android.permission.READ_MEDIA_AUDIO" />
-    <uses-permission android:name="android.permission.INTERNET" />
-
-    <!--If I am not giving this permission, app is crashing saying that it needs this permission-->
+    <uses-permission android:name="android.permission.INTERNET" /> <!-- If I am not giving this permission, app is crashing saying that it needs this permission -->
     <uses-permission android:name="android.permission.FOREGROUND_SERVICE" />
 
     <application
@@ -23,17 +21,24 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.ScrutinizingTheService"
-        tools:targetApi="31">
+        tools:targetApi="31" >
         <activity
-            android:name=".ui.music.MusicPlayerActivity"
-            android:exported="false">
+            android:name=".ui.job_intent.JobIntentActivity"
+            android:exported="false" >
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".ui.NotificationActivity"
-            android:exported="false">
+            android:name=".ui.music.MusicPlayerActivity"
+            android:exported="false" >
+            <meta-data
+                android:name="android.app.lib_name"
+                android:value="" />
+        </activity>
+        <activity
+            android:name=".ui.notif.NotificationActivity"
+            android:exported="false" >
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
@@ -57,7 +62,7 @@
 
         <receiver
             android:name=".broadcastReceivers.MediaBroadcastReceiver"
-            android:exported="false">
+            android:exported="false" >
             <intent-filter>
                 <action android:name="EMITTER_PLAY" />
                 <action android:name="EMITTER_PAUSE" />
@@ -65,10 +70,9 @@
                 <action android:name="EMITTER_PREVIOUS" />
             </intent-filter>
         </receiver>
-
         <receiver
             android:name=".services.MusicPlayerService$MusicPlayerListener"
-            android:exported="false">
+            android:exported="false" >
             <intent-filter>
                 <action android:name="RECEIVER_PLAY" />
             </intent-filter>
@@ -90,8 +94,8 @@
         </receiver>
 
         <activity
-            android:name=".ui.MainActivity"
-            android:exported="true">
+            android:name=".ui.launcher.MainActivity"
+            android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
