+++ app/src/main/java/com/example/scrutinizing_the_service/ui/job_intent/JobIntentActivity.kt
@@ -1,0 +1,11 @@
+package com.example.scrutinizing_the_service.ui.job_intent
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+
+class JobIntentActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_job_intent)
+    }
+}
