+++ app/src/main/res/layout/activity_notification.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ui.NotificationActivity">
+    tools:context=".ui.notif.NotificationActivity">
 
     <Button
         android:id="@+id/btn_show_notif"
